\name{runExample}
\alias{runExample}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Run example \code{evaluate.pm} analyses
}
\description{
Generate random cyclic time series, analyse them using \code{\link{evaluate.pm}} and plot the results.
}
\usage{
runExample()
}
\references{
Louca, S., Doebeli, M. - Detecting cyclicity in ecological time series (in review, as of June 2014)
}
\author{
Stilianos Louca
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
\code{\link{evaluate.pm}}, \code{\link{plotReport}}, \code{\link{generate_ouss}}
}
\examples{
  runExample();
}

