\name{runExample}
\alias{runExample}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Run example \code{evaluate.pm} analyses
}
\description{
Generate random cyclic time series, analyse them using \code{\link{evaluate.pm}} and plot the results.
}
\usage{
runExample()
}
\references{
Louca, S., Doebeli, M. (2015) Detecting cyclicity in ecological time series, Ecology 96: 1724--1732
}
\author{
Stilianos Louca
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
\code{\link{evaluate.pm}}, \code{\link{plotReport}}, \code{\link{generate_ouss}}
}
\examples{
  # this might take a few seconds
  runExample();
}

